added getSlotByName() method to Intent
